# Making vertex optimization 50x faster