… tests/particles/fixtures/particle_test.png