Fixes #273 

Reference benchmark data on my machine:
- BD-5J fuselage: `collectXPlaneObjects` takes 0.6 seconds non-optimized, 20 seconds optimized

New performance:
0.6 seconds (unaffected) non-optimized, 0.38 seconds optimized

The only thing changed is the way to keep track of existing vertices, and the duplicates check. The new method does not incur and overhead, and the raw performance _gain_ from running optimized is now simply from having to deal with a smaller collection of verts (18k vs 44k).

Originally, I had done a few other optimizations in this method. Mainly taking out all global variable reference out of the most critical loop (`for f in tempfaces`). However, it complicated the code (needing local variables, needing to reconcile with the global ones at the end of the loop etc.) and only took 0.1 second off. The tradeoff for maintainability wasn't worth it.