@tngreene 2019-08-19
2 parts:

1. How is this different from current work on the subject. Obviously the code is shorter
2. Can you merge into the branch v3-5 at the tag `v3.5.1-beta.1` (actually could you do this for all your pull requests. Sorry this isn't written down somewhere. It will be)